+++ src/main/java/com/datasec/Server.java
@@ -1,5 +1,4 @@
 package com.datasec;
 
 public class Server {
-    
 }
